fix(front_matter): Move exports to their own module
