+++ front_matter/json.ts
@@ -1,7 +1,10 @@
 // Copyright 2018-2023 the Deno authors. All rights reserved. MIT license.
 
-import { createExtractor, Format, Parser, test as _test } from "./mod.ts";
-export { Format } from "./mod.ts";
+import { createExtractor, Parser } from "./createExtractor.ts";
+import { Format } from "./formats.ts";
+import { test as _test } from "./test.ts";
+
+export { Format } from "./formats.ts";
 
 export function test(str: string): boolean {
   return _test(str, [Format.JSON]);
